# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Neosofia Notification Service API",
    "description": "Relays public contact-form mail to the configured inbox and protected platform mail to caller-approved destinations via Resend.",
    "version": "0.1.0",
    "contact": {
      "name": "Neosofia"
    },
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8005",
      "description": "Local development"
    },
    {
      "url": "https://notification.neosofia.tech",
      "description": "Production"
    }
  ],
  "tags": [
    {
      "name": "Email",
      "description": "Outbound email relay"
    },
    {
      "name": "Health",
      "description": "Service liveness probe"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Liveness probe",
        "description": "Returns 200 when the service is running. Exempt from rate limiting.",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails": {
      "post": {
        "tags": ["Email"],
        "summary": "Relay a contact email",
        "description": "Validates the request, then relays an email to the configured destination inbox via Resend. The caller's address is set as `reply_to` so replies go directly to the sender.\n\n**Rate limits (per originating IP):** 5 per minute · 20 per hour · 30 per hour (global) · 200 per day (global).",
        "operationId": "sendEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              },
              "example": {
                "from_email": "visitor@example.com",
                "subject": "General inquiry",
                "message": "Hello, I would like to learn more about your services."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email accepted and relayed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailResponse"
                },
                "example": {
                  "status": "sent"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — one or more required fields are missing or empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "from_email is required; message is required"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "5 per 1 minute"
                }
              }
            }
          },
          "502": {
            "description": "Upstream Resend API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Failed to relay message. Please try again later."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails": {
      "post": {
        "tags": ["Email"],
        "summary": "Relay a protected platform email",
        "description": "Requires a valid platform service JWT, validates the destination against the configured allowlist, and relays the message via Resend.\n\n**Rate limits (per authenticated subject):** 30 per minute · 120 per hour.",
        "operationId": "sendPlatformEmail",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformEmailRequest"
              },
              "example": {
                "to_email": "alerts@example.com",
                "subject": "Escalation",
                "message": "Open the dashboard for details.",
                "from_email": "workflow@example.com",
                "reply_to": "oncall@example.com",
                "message_type": "clinical-alert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email accepted and relayed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailResponse"
                },
                "example": {
                  "status": "sent"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — malformed request or disallowed destination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "JWT subject is not permitted to call the relay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream Resend API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Protected relay is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "EmailRequest": {
        "type": "object",
        "required": ["from_email", "subject", "message"],
        "properties": {
          "from_email": {
            "type": "string",
            "format": "email",
            "description": "Sender's email address; set as reply_to on the outbound email.",
            "example": "visitor@example.com"
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Email subject. Prefixed with [Contact] on delivery. Must not contain control characters.",
            "example": "General inquiry"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Plain-text message body.",
            "example": "Hello, I would like to learn more about your services."
          }
        }
      },
      "PlatformEmailRequest": {
        "type": "object",
        "required": ["to_email", "subject", "message"],
        "properties": {
          "to_email": {
            "type": "string",
            "format": "email",
            "description": "Destination inbox for the protected relay. The domain must be allowlisted by PLATFORM_EMAIL_ALLOWED_DOMAINS.",
            "example": "alerts@example.com"
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Email subject. Prefixed with the message_type on delivery. Must not contain control characters.",
            "example": "Escalation"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Plain-text message body.",
            "example": "Open the dashboard for details."
          },
          "from_email": {
            "type": "string",
            "format": "email",
            "description": "Optional sender email included in the relayed message body."
          },
          "reply_to": {
            "type": "string",
            "format": "email",
            "description": "Optional reply-to address applied to the outbound email."
          },
          "message_type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9_-]+$",
            "description": "Optional delivery and logging classifier used as the subject prefix.",
            "default": "platform",
            "example": "clinical-alert"
          }
        }
      },
      "EmailResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["sent"],
            "description": "Confirmation that the email was accepted by Resend."
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": ["status", "version"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"]
          },
          "version": {
            "type": "string",
            "description": "Service semver from pyproject at install/build time."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      }
    }
  }
}